[CELEBORN-1486][FOLLOWUP] Update document link of Get Started With Velox and Get Started With ClickHouse in glutensupport.md
### What changes were proposed in this pull request?

Update document link of `Get Started With Velox` and `Get Started With ClickHouse` in `glutensupport.md`. Meanwhile, replace `gluten-celeborn-package-xx-SNAPSHOT.jar` with `(The bundled Gluten Jar. Make sure -Pceleborn is specified when it is built.)`, which refers to https://github.com/apache/incubator-gluten/pull/6692.

### Why are the changes needed?

The document link of `Get Started With Velox` and `Get Started With ClickHouse` could not access, which has already changed the url.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

No.

Closes #2762 from SteNicholas/CELEBORN-1486.

Authored-by: SteNicholas <programgeek@163.com>
Signed-off-by: mingji <fengmingxiao.fmx@alibaba-inc.com>